+++ db/migrate/20130802220629_remove_project_emails.rb
@@ -1,0 +1,5 @@
+class RemoveProjectEmails < ActiveRecord::Migration
+  def change
+    remove_column :projects, :emails, :text
+  end
+end