+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Gem Changes**
   - Updated to contour 2.1.0.beta2
+- **Project Changes**
+  - Project emails field was removed since it is no longer used
 
 ## 0.19.0 (August 1, 2013)